# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Plaid+Stripe API Documentation",
        "description": "API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/stripe/check-payment-intent": {
            "post": {
                "tags": [
                    "Stripe Status Check"
                ],
                "summary": "Check the status of a Stripe payment intent",
                "description": "Returns the status of a Stripe payment intent",
                "operationId": "c7bc9ac74bf862fb37036a5870d16b7d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "payment_intent_id": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        }
    },
...
}